DoctrineKeyValueStyleRule: stricter type checks
We now require that column types _do_ accept the value type, whereas
previously we only required that column types _maybe_ accept the value
type.

This fixes some cases where errors were not properly being reported,
such as if the value type is nullable, but the column has a "NOT NULL"
constraint.

The one exception we make is that integer ranges are still not checked,
because it is likely uncommon for codebases to have range restrictions
on all the integers that are passed to the database. Checking integer
ranges in this rule could maybe be part of a future "strict mode".