DoctrineKeyValueStyleRule: properly handle unions
If any of the arguments (table name or parameter values) to the
DoctrineKeyValueStyleRule methods are a union type, each subtype will
now be checked independently.

Previously, union types would cause an error, because of the improper
use of `instanceof *Type` checks. This is now fixed by using methods
to get the constants, e.g. `getConstantStrings()`, as recommended in
https://github.com/staabm/phpstan-dba/pull/564#discussion_r1133053061.