Revert back to isSuperTypeOf
Handle the `FloatType` explicitly instead of testing with `accepts`,
since that function does more than what we want.

Note that this makes benevolent union types unacceptable again unless
all types in the union are acceptable.

I've also simplified the logic for casting an `IntegerRangeType` into
an `IntegerType`, since a column type can only reasonable have a union
with a `NullType`. There is a standard way to using TypeCombinator to
mutate the type of a union with null.