SQL AST: narrow result types with WHERE condition
The WHERE condition of a SQL statement may narrow the type returned
from a table. We can account for this when getting the type of each
selected column.

Currently, the only implemented conditions are `IS NOT NULL` and
`IS NULL`. This can be expanded in the future to support other binary
operators and comparison operators.

Since the WHERE condition can narrow selections that are implicitly
specified with an asterisk, we now expand the asterisk to an explicit
list of columns so that we can perform the AST-based type resolution
on them.